+++ docs/03_guides/05_crawlee.mdx
@@ -5,9 +5,9 @@ title: Using Crawlee
 
 import CodeBlock from '@theme/CodeBlock';
 
-import CrawleeBeautifulSoupExample from '!!raw-loader!./code/02_crawlee_beautifulsoup.py';
-import CrawleeParselExample from '!!raw-loader!./code/02_crawlee_parsel.py';
-import CrawleePlaywrightExample from '!!raw-loader!./code/02_crawlee_playwright.py';
+import CrawleeBeautifulSoupExample from '!!raw-loader!./code/05_crawlee_beautifulsoup.py';
+import CrawleeParselExample from '!!raw-loader!./code/05_crawlee_parsel.py';
+import CrawleePlaywrightExample from '!!raw-loader!./code/05_crawlee_playwright.py';
 
 In this guide you'll learn how to use the [Crawlee](https://crawlee.dev/python) library in your Apify Actors.
