+++ docs/03_guides/code/04_selenium.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import asyncio
 from urllib.parse import urljoin
